-Change API requester TransactionInquiry.kt -Add Read Connection TImeout for QR API Callback response -Add New API VoidPayment -Add Swipe To Delete RecycleView

+++ app/src/main/java/com/example/mockingmerchantapp/ModelClass/TransactionInquiry.kt
@@ -2,6 +2,7 @@ package com.example.mockingmerchantapp.ModelClass
 
 import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
+import java.math.BigDecimal
 
 data class TransactionInquiry(
     @Expose
@@ -33,7 +34,7 @@ data class Transactions(
     val when_: String,
     @Expose
     @SerializedName("amount")
-    var amount: Int,
+    var amount: Double,
     @Expose
     @SerializedName("status")
     val status: String,
